+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/impl/AbstractPolyglotImpl.java
@@ -1139,6 +1139,12 @@ public abstract class AbstractPolyglotImpl {
 
         public abstract boolean removeMember(Object context, Object receiver, String key);
 
+        public boolean hasStaticReceiver(Object context, Object receiver) {
+            return false;
+        }
+
+        public abstract Object getStaticReceiver(Object context, Object receiver);
+
         public boolean canExecute(Object context, Object receiver) {
             return false;
         }