+++ truffle/src/com.oracle.truffle.api.test/src/com/oracle/truffle/api/test/polyglot/ValueHostConversionTest.java
@@ -40,6 +40,7 @@
  */
 package com.oracle.truffle.api.test.polyglot;
 
+import static com.oracle.truffle.tck.tests.ValueAssert.Trait.STATIC_RECEIVER;
 import static com.oracle.truffle.tck.tests.ValueAssert.assertUnsupported;
 import static com.oracle.truffle.tck.tests.ValueAssert.assertValue;
 import static com.oracle.truffle.tck.tests.ValueAssert.Trait.BOOLEAN;
@@ -162,7 +163,7 @@ public class ValueHostConversionTest extends AbstractPolyglotTest {
 
         assertTrue(value.isHostObject());
         assertSame(obj, context.asValue(value.as(Object.class)).asHostObject());
-        assertUnsupported(value, HOST_OBJECT, MEMBERS);
+        assertUnsupported(value, HOST_OBJECT, MEMBERS, STATIC_RECEIVER);
     }
 
     /**
@@ -258,7 +259,7 @@ public class ValueHostConversionTest extends AbstractPolyglotTest {
         assertTrue(newInstance.getMetaObject().newInstance().asHostObject() instanceof JavaRecord);
         assertSame(JavaRecord.class, newInstance.getMetaObject().asHostObject());
 
-        assertValue(recordClass, Trait.INSTANTIABLE, Trait.MEMBERS, Trait.HOST_OBJECT, Trait.META);
+        assertValue(recordClass, Trait.INSTANTIABLE, Trait.MEMBERS, Trait.HOST_OBJECT, Trait.META, STATIC_RECEIVER);
     }
 
     @Test
@@ -316,8 +317,8 @@ public class ValueHostConversionTest extends AbstractPolyglotTest {
         assertTrue(record.hasMember("wait"));
         assertTrue(record.hasMember("notifyAll"));
 
-        assertValue(record, Trait.MEMBERS, Trait.HOST_OBJECT);
-        assertValue(record.getMetaObject(), Trait.INSTANTIABLE, Trait.MEMBERS, Trait.HOST_OBJECT, Trait.META);
+        assertValue(record, Trait.MEMBERS, Trait.HOST_OBJECT, STATIC_RECEIVER);
+        assertValue(record.getMetaObject(), Trait.INSTANTIABLE, Trait.MEMBERS, Trait.HOST_OBJECT, Trait.META, STATIC_RECEIVER);
     }
 
     @Test